fix(work): satisfy cross-batch CI checks
